% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{bigquery_billing_project}
\alias{bigquery_billing_project}
\title{Get the BigQuery billing project ID}
\usage{
bigquery_billing_project()
}
\description{
If bigquery_billing_project isn't set, raises an error.
}
