% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_tags}
\alias{summarize_tags}
\title{Given a grouped remote table of posts, summarize tags}
\usage{
summarize_tags(tbl, tags = NULL, question_id = "id")
}
\arguments{
\item{tbl}{A remote grouped table based on}

\item{tags}{Optionally, a vector of tags to summarize}

\item{question_id}{Which column represents the question ID?}
}
\description{
Given a grouped remote table of posts, summarize tags
}
